chore(benchmark): Increase wait time for UI elements
This commit increases the wait time for UI elements in macrobenchmark tests from 5 seconds to 10 seconds.

This change aims to improve the reliability of the benchmarks by allowing more time for the UI to settle before performing actions.

Additionally, the `mainScrollNoteDownUp` function has been commented out as it is not currently being used in the benchmarks.